requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Django==3.1.1
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
jieba==0.42.1
kiwisolver==1.2.0
line-bot-sdk==1.17.0
lxml==4.5.2
matplotlib==3.3.2
numpy==1.19.1
openpyxl==3.0.5
pandas==1.1.2
Pillow==7.2.0
pymssql==2.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
twder==0.1.4
urllib3==1.25.10
